fix: fixed header issues and reworked directory
