+++ app/page.tsx
@@ -1,0 +1,9 @@
+const HomePage = () => {
+  return (
+    <div>
+      <h1>Hi there</h1>
+    </div>
+  )
+}
+
+export default HomePage
