+++ lib/helper.ts
