--- src/app/head.tsx
@@ -1,11 +1,0 @@
-import Seo from '@/components/Seo'
-
-const Header = () => {
-  return (
-    <>
-      <Seo />
-    </>
-  )
-}
-
-export default Header
