--- src/app/page.tsx
@@ -1,12 +1,0 @@
-import Layout from '@/app/layout'
-
-const HomePage = () => {
-  return (
-    <Layout>
-      {/* <Seo templateTitle='Home' /> */}
-      <h1>Hi there</h1>
-    </Layout>
-  )
-}
-
-export default HomePage
