--- src/pages/_app.tsx
@@ -1,15 +1,0 @@
-import { AppProps } from 'next/app'
-
-import '@/styles/globals.css'
-// import '@/styles/colors.css';
-
-/**
- *  info
- * ? `Layout` component is called in every page using `np` snippets. If you have consistent layout across all page, you can add it here too
- */
-
-function MyApp({ Component, pageProps }: AppProps) {
-  return <Component {...pageProps} />
-}
-
-export default MyApp
