--- src/pages/_document.tsx
@@ -1,21 +1,0 @@
-import { Head, Html, Main, NextScript } from 'next/document'
-
-export default function Document() {
-  return (
-    <Html lang='en'>
-      <Head>
-        <link
-          rel='preload'
-          href='/fonts/inter-var-latin.woff2'
-          as='font'
-          type='font/woff2'
-          crossOrigin='anonymous'
-        />
-      </Head>
-      <body>
-        <Main />
-        <NextScript />
-      </body>
-    </Html>
-  )
-}
